// src/types.ts
export type TUserRegistrationInfo = {
    name: string;
    password: string;
    phoneNumber: string;
    companyName: string;
    isBoss: boolean;
}

export type TUserLoginInfo = {
    name: 'string';
    password: 'string';
}

export type TCompanyInfo = {
    companyName: string;
    country: string;
}